PublicationUpload now belongs to the toolbar
